+++ client_app/src/components/shelter/ScheduleData.js
@@ -1,0 +1,23 @@
+// client_app/src/components/shelter/ScheduleData.js
+export const ScheduleData = {
+  Content: [
+    {
+      name: "Morning Shift",
+      start: 8 * 3600000,   // 8:00 AM
+      end: 12 * 3600000,    // 12:00 PM
+      people: 5,
+    },
+    {
+      name: "Afternoon Shift",
+      start: 13 * 3600000,  // 1:00 PM
+      end: 17 * 3600000,    // 5:00 PM
+      people: 4,
+    },
+    {
+      name: "Evening Shift",
+      start: 18 * 3600000,  // 6:00 PM
+      end: 22 * 3600000,    // 10:00 PM
+      people: 3,
+    },
+  ],
+};
