--- client_app/src/components/shelter/ScheduleData.tsx
@@ -1,24 +1,0 @@
-// src/components/shelter/ScheduleData.tsx
-export const ScheduleData = {
-    Content: [
-      {
-        name: "Morning Shift",
-        start: 8 * 3600000,   // 8:00 AM
-        end: 12 * 3600000,    // 12:00 PM
-        people: 5,
-      },
-      {
-        name: "Afternoon Shift",
-        start: 13 * 3600000,  // 1:00 PM
-        end: 17 * 3600000,    // 5:00 PM
-        people: 4,
-      },
-      {
-        name: "Evening Shift",
-        start: 18 * 3600000,  // 6:00 PM
-        end: 22 * 3600000,    // 10:00 PM
-        people: 3,
-      },
-    ],
-  };
-
